rag/requirements.txt
# -c ../constraints.txt
fastapi
uvicorn[standard]
qdrant-client
openai
python-dotenv